requirements.txt
click==7.1.2
Flask==1.1.2
Flask-HTTPAuth==4.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.10.0
psycopg2==2.8.6
PyJWT==2.0.1
Werkzeug==1.0.1
